refactor: move deprecated Button and Dropdown into color-picker
