+++ src/components/color-picker/deprecated-button/deprecated-button.less
@@ -1,4 +1,4 @@
-@import '../styles/constants.less';
+@import '../../styles/constants.less';
 
 .reactist_button {
     cursor: pointer;
@@ -98,7 +98,7 @@
                 animation-name: reactistRotateRight;
                 animation-timing-function: linear;
                 color: #ffffff;
-                background-image: url('../icons/spinner.svg');
+                background-image: url('../../icons/spinner.svg');
             }
         }
     }
