+++ src/components/color-picker/deprecated-button/deprecated-button.test.tsx
