+++ src/components/color-picker/deprecated-button/index.ts
