+++ src/components/color-picker/deprecated-dropdown/dropdown.less
@@ -1,4 +1,4 @@
-@import '../styles/constants.less';
+@import '../../styles/constants.less';
 
 .reactist_dropdown {
     .trigger {
