+++ src/components/color-picker/deprecated-dropdown/dropdown.test.tsx
